Make sure that the child processes from the cluster tests:
- Do not try to rerun all of the subsequent suites.
- Runs the correct version of prom-client in the child processes

This should improve the screen output a good bit, speed up the
tests and make the benchmarks for the sender side of the code
accurate.

benchmark.js and benchmark-regressions are fighting for control
of the benchmark lifecycle and it is inordinately difficult to
thread that needle where workers or clusters are involved.

It would be good if there were an alternative to this mess out
there somewhere, but I know of none.