Add a Map that can do lookups by label sets for configured labels.
By requiring 'sortedLabelNames' to be available, we can generate much
shorter keys for the tables. This should help a great deal with #611

Unfortunately this cannot simply be fixed with changes to hashObject()
because this implementation detail has leaked out across the entire
codebase. So a new implementation that can be subbed into Metrics one
implementation at a time (they all deal with entry.value in a unique
way)